Put lfs user removal after mk_SUDO
Since --from-lfs is used when chown'ing files, and this is done
during the mk_SUDO target, the lfs user should still exist during
this target. Right now, we remove the --from-lfs stanzas
from the scriptlet, but it is much better to move lfs user removal.
This commit does this, and the next commit will revert the
--from-lfs removal.