Change the {save,restore}-luser targets
As discussed in ticket #1739, we don't want to remove the lfs user
anymore. This means we don't need to restore it afterwards either.
So if an lfs user exists, it has to be renamed to something else
(lfsxxx), and the lfs user will be (re)created by book instructions.
The save-luser target does this. The restore-luser target is just
a warning about the states of the lfs* users.