New management of luser and lgroup
The new management is the following:
- the jhalfs user should choose a user name for the user running
  commands in chapter 5 and 6
- this user is created with chapter 4 instructions, and removed
  before entering chroot
We don't need to save an existing user nor to restore it after
the jhalfs run, so that we remove the {save,restore}-luser targets.
We introduce a check-luser target that just verifies that the chosen
name does not exist, and errors out if it exists.
We introduce a remove-luser target, that silently removes the
temporary user, and is called before entering chroot