Output preceding linefeed only for <screen>
See ticket #1743 for details. Because of the book layout, it is
necessary to output a linefeed between two blocks of commands.
Previously, this was done before each <userinput> tag. This
is wrong, since <userinput> can occur several times inside
a block. We now only output linefeed before <screen> or
before <command> when that one contains a test instruction.

Fixes #1743