Check that $(MOUNT_PT) is a mount point
See ticket #1742 for details. It has been decided to add a check
at the beginning of the run of make in the jhalfs directory.
If $(MOUNT_PT) is not a mount point, which may be a valid use case,
we don't stop, but we warn the user and give her/him 5 seconds for
stopping the process. If the <return> key is not hit within 5
seconds, the build starts.
This target is a dependency of the "all" Makefile target, so that
it is run each time make is started without parameter...

Fixes #1742.