Make extraction of packages from compound more robust
When we want to build a package from a compound one, we look for
' '<package>-, but this may match ' '<package>-<something>', either
in the same compound package or even in another one. So we have to
test that the first character after the dash is a digit. Note that
we also modify the search for the tarball (remove recursive calls).
This is a slight optimization. Then the tarball name is verified to
contain <package>-<digit>. If not the tarball search is recursively
called for the following occurence. Note that if the compound
package is not the right one, the returned tarball is empty.