BLFS: matching tarball of compound: return full line
The matching tarball algorithm only returned the tarball name
and did not test whether the line was commented out. So add this
test, and return the full line: this will allow using simple
string functions for the md5 and the download directory
instead of matching again the tarball...
We change the name of the template to a more explicit one.
Next change will use the line in the compound sect1 template.