BLFS: set CMAKE_ALWAYS_INSTALL if porg is used
cmake does not reinstall files that are already installed and not
modified. So not all the files are logged by porg, and the package
information is not accurate. The variable CMAKE_ALWAYS_INSTALL has
to be set in the environment to force cmake to install all the
files. We set it even if cmake is not used, which of course does
nothing. But detecting whther a package uses cmake, although doable,
is cumbersome...

Fixes #1719