Add TEST_MISMATCH set to the opposite of GETPKG
This is for checking a version mismatch between book and directory name.
The preceding commits allow to do this check when downloading packages,
but if "Retrieve source files" is not ticked, this is not done. In this
case we do it in scriptlets. But we have to pass a parameter to the
xsl stylesheet to tell to generate this test in scriptlet.

This commit defines a variable TEST_MISMATCH set to the opposite of GETPKG.

Part of #1740