Generate a test in the scriptlets
This is for checking a version mismatch between book and directory name.
The preceding commits allow to do this check when downloading packages,
but if "Retrieve source files" is not ticked, this is not done. In this
case we do it in scriptlets. But we have to pass a parameter to the
xsl stylesheet to tell to generate this test in scriptlet.

This commit adds the parameter test-mismatch to the xsl stylesheet, and
some code to generate the test in the scriptlet if test-mismatch is 'y'.

Fixes #1740