BLFS: fix "time after make" coming after install
When the user asks to generate statistics for a package, tests are run
with some statistics. If tests come just after make, we need to include
"time after make" and "size after make", but if tests come after
installing to the system, this timing is not necessary (it has been
done after the destdir install). So don't generate it in this case.