Comment out "cat vim-test.log" when tests disabled
We recently added a line after vim tests to copy logs to $TEST_LOG.
Problem is that vim-tests.log is not created if the tests are not run
but the line is never commented out. This leads to an error...
The idea is just to call the template again with
"cat vim-tests.log". This will comment out the line when needed,
always redirect vim-tests.log and any error to $TEST_LOGS,
and not stop on error.

Fix bug reported in
https://lists.linuxfromscratch.org/sympa/arc/alfs-discuss/2025-02/msg00011.html