blfs-tools: process all role="installation" sect2
Previously, in scripts.xsl, we had a case for role="installation"
and not preceding-siblings::sect2[@role="installation"]. But
no case when there was a preceding sect2 with role="installation".
We remove the condition that there is such a preceding sect2, and
replace with a "if" that allows the first such sect2 to call a
template that generates the unpack code.
When logging size/time information, log also the title of the
sect2.